Revision 3.5
- Added the InvoiceReport.java class

- Made various edits to Invoice.java to correspond properly with the
InvoiceReports class

- Modified the Rental.java class

- Still a few errors, and likely a lot more fixing needing to be done in
the future, however everything should still function correctly.

+++ src/com/vgb/Rental.java
@@ -18,7 +18,7 @@ public class Rental extends Item {
 	 * @param rentalCost  The base cost (price) of the equipment.
 	 * @param hours       The number of rental hours.
 	 */
-	public Rental(UUID uuid, char type, String name, String model, int rentalCost, double hours) {
+	public Rental(UUID uuid, char type, String name, String model, double rentalCost, double hours) {
 		super(uuid, type, name);
 		this.model = model;
 		this.rentalCost = rentalCost;
